minor-adon: mbapy\sci_instrument\_base.py: SciInstrumentData: process_raw_data: add y_scale parameter for scaling y-axis data
Introduce a `y_scale` parameter in the `process_raw_data` method of `SciInstrumentData` class to allow for calibration of the y-axis data. This feature is particularly useful when the data needs to be scaled before further processing or analysis.
Additionally, updated the import order in `SCIEX.py` for consistency and applied a minor code cleanup.