feat(bio): add MutationOpts and MutationTree classes for peptide mutations
Implement new classes MutationOpts and MutationTree to facilitate peptide
mutation operations. This extends the functionality of the Peptide class by
allowing for more complex manipulations and optimizations in the mutation
process. The classes provide options for AA deletion, repetition, and
replacement, as well as management of N/C/R-protect group deletions.